# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Games4King Workplace OS API
  version: 1.0.0
  description: Internal employee portal API for Games4King
  license:
    name: MIT
servers:
  - url: /api
    description: API proxy (via Next.js rewrites)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
...
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    get:
      tags:
        - Auth
      summary: Refresh access token via HttpOnly refresh cookie
      security: []
      responses:
        '200':
          description: New access token
        '401':
          description: Invalid or expired refresh token
...